+++ src/globals.js
@@ -1,0 +1,23 @@
+/**
+ * JavaScript Promises (ES6)
+ * @external Promise
+ * @see http://www.html5rocks.com/en/tutorials/es6/promises/
+ */
+
+/**
+ * HTML `<canvas>` element
+ * @external canvas
+ * @see https://developer.mozilla.org/en-US/docs/Web/HTML/Element/canvas
+ */
+
+/**
+ * A `THREE.Texture` object instance
+ * @external Texture
+ * @see http://threejs.org/docs/#Reference/Textures/Texture
+ */
+
+/**
+ * CanvasRenderingContext2D for drawing onto a `<canvas>`
+ * @external CanvasRenderingContext2D
+ * @see https://developer.mozilla.org/en-US/docs/Web/API/CanvasRenderingContext2D
+ */