+++ src/texture-manager.js
@@ -1,16 +1,59 @@
-/*
- * Manages the texture canvas(es) for the system labels using the Knapsack class
+/**
+Build and destroy "nodes" in your texture atlas easily. It builds one or more {@link module:texture-manager/knapsack|`Knapsack`} objects for you, each of which represent a separate square texture atlas with one or more sprite textures of a size defined by you.
+
+@module texture-manager
+
+@example
+  // ES2015 modules
+  import TextureManager from 'three-sprite-texture-atlas-manager';
+  var textureManager = new TextureManager();
+
+@example
+  // node.js, requirejs
+  var TextureManager = require('three-sprite-texture-atlas-manager');
+  var textureManager = new TextureManager();
+
+@example
+  // global
+  var textureManager = new window.threeSpriteAtlasTextureManager();
+ *
  */
 
-import Knapsack from './knapsack';
+import Knapsack from './texture-manager/knapsack';
 
+/**
+  * @constructor
+  * @param {integer} [size=1024] Optional size for the textures. Must be a power of two.
+  */
 class TextureManager {
   constructor( size ) {
-    this.size = ( typeof size === 'number' && /^(128|256|512|1024|2048|4096|8192|16384)$/.test( size ) ) ? size : 1024;
+    /**
+     * The size of the textures as validated after constructing the object.
+     * @member {integer} size
+     * @readonly
+     */
+    this.size = ( ( typeof size === 'number' ) && /^(128|256|512|1024|2048|4096|8192|16384)$/.test( size ) ) ? size : 1024;
+
+    /**
+     * As the texture manager allocates nodes, it creates a new {@link module:texture-manager/knapsack|`Knapsack`} when it needs to provide space for nodes. This is an array with all the knapsacks which have been created.
+     * @member {Knapsack[]} knapsacks
+     * @readonly
+     */
     this.knapsacks = [];
+
+    /**
+     * The debug property can be set to `true` after instantiating the object, which will make the {@link module:texture-manager/knapsack/node|`KnapsackNode`} class draw outlines as it allocates nodes. This can make it much more obvious what is going on, such as whether your text is properly sized and centered.
+     * @member {boolean} debug
+     */
     this.debug = false;
   }
 
+  /**
+   * Add a new knapsack to the texture manager.
+   * @param {integer} size
+   * @returns {Knapsack}
+   * @private
+   */
   _addKnapsack( size ) {
     var knapsack = new Knapsack( this, size );
     this.knapsacks.push( knapsack );
@@ -20,13 +63,21 @@ class TextureManager {
     return knapsack;
   }
 
+  /**
+   * The size of the texture
+   * @type {integer}
+   * @readonly
+   */
   get textureSize () {
     return this.size;
   }
 
-  // Claim a texture atlas slot for an image of width x height pixels,
-  // will create a new texture atlas if needed.
-  // Returns a Promise.
+  /**
+   * Allocate a texture atlas node for a sprite image of `width` by `height` pixels.
+   * @param {integer} width
+   * @param {integer} height
+   * @returns {external:Promise}
+   */
   allocateNode( width, height ) {
     var self = this;
 
@@ -73,6 +124,10 @@ class TextureManager {
     }));
   }
 
+  /**
+   * Release the given node.
+   * @param {KnapsackNode} node
+   */
   release( node ) {
     if ( node ) {
       node.release();
@@ -81,3 +136,6 @@ class TextureManager {
 }
 
 export default TextureManager;
+
+/**
+ */