+++ CHANGELOG.md
@@ -1,3 +1,15 @@
 ### [0.0.1](https://github.com/Leeft/three-sprite-texture-atlas-manager/releases/tag/v0.0.1)
 
 - The first release
+
+### [0.0.2](https://github.com/Leeft/three-sprite-texture-atlas-manager/releases/tag/v0.0.2)
+
+- Build/release process/documentation fixes
+
+### [0.0.3](https://github.com/Leeft/three-sprite-texture-atlas-manager/releases/tag/v0.0.3)
+
+- Build/release process/documentation fixes
+
+### [0.0.4](https://github.com/Leeft/three-sprite-texture-atlas-manager/releases/tag/v0.0.4)
+
+- Build/release process/documentation fixes